Split account API components into separate modules
